Deploy gui

+++ ExtractFeature/sniff/sniffer.go
@@ -1,15 +1,16 @@
 package sniff
 
 import (
+	"FlowDetection/GUI"
 	"FlowDetection/baseUtil"
 	"FlowDetection/flowFeature"
+	"errors"
 	"fmt"
 	"log"
 	"os"
 	"time"
 
 	"github.com/google/gopacket"
-	"github.com/google/gopacket/ip4defrag"
 	"github.com/google/gopacket/layers"
 	"github.com/google/gopacket/pcap"
 	"github.com/google/gopacket/pcapgo"
@@ -20,16 +21,45 @@ const (
 	timeout            = 1 * time.Second
 )
 
+var (
+	BlackToSnifferChan chan *GUI.OperateSniffer
+	BlackList          map[string]interface{}
+	WriteFile          *bool
+	wPcap              *pcapgo.Writer
+)
+
+func init() {
+	BlackList = make(map[string]interface{}, 100)
+}
+
+func receiveBlack() {
+	for {
+		select {
+		case black, ok := <-BlackToSnifferChan:
+			if !ok {
+				log.Fatal("黑名单接收信道出现问题")
+			}
+
+			_, ok = BlackList[black.IP]
+			if !ok && black.Operate == 1 {
+				BlackList[black.IP] = struct{}{}
+			} else if ok && black.Operate == 0 {
+				delete(BlackList, black.IP)
+			}
+		}
+	}
+}
+
 //usage:
 //	NewSniffer()
 //	SetSnifferInterface(device string)
 //	StartSniffer()
 type Sniffer struct {
-	Devices          []pcap.Interface
-	handle           *pcap.Handle
-	FragmentList     *ip4defrag.IPv4Defragmenter
+	Devices []pcap.Interface
+	handle  *pcap.Handle
+	packets chan gopacket.Packet
+
 	conversationPool *ConversationPool
-	connMsg          map[uint16]*ConnMsg
 }
 
 func NewSniffer(featureChan chan *flowFeature.FlowFeature) (*Sniffer, error) {
@@ -39,9 +69,7 @@ func NewSniffer(featureChan chan *flowFeature.FlowFeature) (*Sniffer, error) {
 	}
 	return &Sniffer{
 		Devices:          devices,
-		FragmentList:     ip4defrag.NewIPv4Defragmenter(),
 		conversationPool: NewConversationPool(featureChan),
-		connMsg:          make(map[uint16]*ConnMsg),
 	}, nil
 }
 
@@ -58,8 +86,35 @@ func (sniffer Sniffer) PrintDevices() {
 	}
 }
 
+//sourceType:     1表示网卡， 0表示pcap文件
+func (sniffer *Sniffer) SetSnifferSource(sourceName string,
+	sourceType uint, promiscuous bool) error {
+
+	//如果是嗅探网卡
+	if sourceType == 1 {
+		err := sniffer.setSnifferInterface(sourceName, promiscuous)
+		if err != nil {
+			return err
+		}
+
+		packetSource := gopacket.NewPacketSource(sniffer.handle, sniffer.handle.LinkType())
+		sniffer.packets = packetSource.Packets()
+	} else if sourceType == 0 { //如果是分析文件
+		err := sniffer.setSnifferFile(sourceName)
+		if err != nil {
+			return err
+		}
+		packetSource := gopacket.NewPacketSource(sniffer.handle, sniffer.handle.LinkType())
+		sniffer.packets = packetSource.Packets()
+	} else {
+		return errors.New("sniffer源类型异常")
+	}
+
+	return nil
+}
+
 //设置嗅探网卡
-func (sniffer *Sniffer) SetSnifferInterface(device string, promiscuous bool) error {
+func (sniffer *Sniffer) setSnifferInterface(device string, promiscuous bool) error {
 	handle, err := pcap.OpenLive(device, int32(snapshotLen), promiscuous, timeout)
 	if err != nil {
 		return err
@@ -70,45 +125,53 @@ func (sniffer *Sniffer) SetSnifferInterface(device string, promiscuous bool) err
 
 }
 
-type ConnMsg struct {
-	srcIP, dstIP [4]byte
-	Start        time.Time
-	Last         time.Time
-	wrong        int
+//分析文件
+func (sniffer *Sniffer) setSnifferFile(fileName string) (err error) {
+	sniffer.handle, err = pcap.OpenOffline(fileName)
+	if err != nil {
+		return
+	}
+	return
 }
 
-func (sniffer *Sniffer) StartSniffer() {
+func (sniffer *Sniffer) StartSniffer(blackToSnifferChan chan *GUI.OperateSniffer, writeFile *bool) {
 	defer sniffer.handle.Close()
 
+	go receiveBlack()
 	go sniffer.conversationPool.checkResultChan()
 
-	//写PCAP文件
-	if baseUtil.CheckFileIsExist("test.pcap") {
-		_ = os.Remove("test.pcap")
-	}
-	f, _ := os.Create("test.pcap")
-	w := pcapgo.NewWriter(f)
-	_ = w.WriteFileHeader(snapshotLen, layers.LinkTypeEthernet)
-	defer f.Close()
+	BlackToSnifferChan = blackToSnifferChan
+	WriteFile = writeFile
 
-	packetSource := gopacket.NewPacketSource(sniffer.handle, sniffer.handle.LinkType())
-	packets := packetSource.Packets()
+	//写PCAP文件
+	if *WriteFile {
+		if baseUtil.CheckFileIsExist("test.pcap") {
+			_ = os.Remove("test.pcap")
+		}
+		f, _ := os.Create("test.pcap")
+		wPcap = pcapgo.NewWriter(f)
+		_ = wPcap.WriteFileHeader(snapshotLen, layers.LinkTypeEthernet)
+		defer f.Close()
+	}
 
 	for {
 		select {
-		case packet := <-packets:
-
-			err := w.WritePacket(packet.Metadata().CaptureInfo, packet.Data())
-			if err != nil {
-				log.Println(err)
+		case packet := <-sniffer.packets:
+			//写PCAP文件
+			if *WriteFile {
+				err := wPcap.WritePacket(packet.Metadata().CaptureInfo, packet.Data())
+				if err != nil {
+					log.Println(err)
+				}
 			}
 
 			sniffer.conversationPool.DisposePacket(packet)
 		case <-time.After(2 * time.Second):
 			log.Println("2秒内没有连接到达...(sniffer.go 112)")
+			sniffer.conversationPool.checkTimeout(time.Now())
+
 		}
 
 	}
 
 }
-
